Struct futures_locks::Mutex

pub struct Mutex<T: ?Sized> { /* fields omitted */ }

A Futures-aware Mutex.

std::sync::Mutex cannot be used in an asynchronous environment like Tokio, because a mutex acquisition can block an entire reactor. This class can be used instead. It functions much like std::sync::Mutex. Unlike that class, it also has a builtin Arc, making it accessible from multiple threads. It's also safe to clone. Also unlike std::sync::Mutex, this class does not detect lock poisoning.

Examples

let mtx = Mutex::<u32>::new(0);
let fut = mtx.lock().map(|mut guard| { *guard += 5; });
spawn(fut).wait_future();
assert_eq!(mtx.try_unwrap().unwrap(), 5);

Methods

impl<T> Mutex<T>

pub fn new(t: T) -> Mutex<T>

Create a new Mutex in the unlocked state.

pub fn try_unwrap(self) -> Result<T, Mutex<T>>

Consumes the Mutex and returns the wrapped data. If the Mutex still has multiple references (not necessarily locked), returns a copy of self instead.

impl<T: ?Sized> Mutex<T>

pub fn get_mut(&mut self) -> Option<&mut T>

Returns a reference to the underlying data, if there are no other clones of the Mutex.

Since this call borrows the Mutex mutably, no actual locking takes place -- the mutable borrow statically guarantees no locks exist. However, if the Mutex has already been cloned, then None will be returned instead.

Examples

let mut mtx = Mutex::<u32>::new(0);
*mtx.get_mut().unwrap() += 5;
assert_eq!(mtx.try_unwrap().unwrap(), 5);

pub fn lock(&self) -> MutexFut<T>

Acquires a Mutex, blocking the task in the meantime. When the returned Future is ready, this task will have sole access to the protected data.

pub fn try_lock(&self) -> Result<MutexGuard<T>, ()>

Attempts to acquire the lock.

If the operation would block, returns Err instead. Otherwise, returns a guard (not a Future).

Examples

let mut mtx = Mutex::<u32>::new(0);
match mtx.try_lock() {
    Ok(mut guard) => *guard += 5,
    Err(()) => println!("Better luck next time!")
};

impl<T: 'static + ?Sized> Mutex<T>

pub fn with<F, B, R, E>(&self, f: F) -> Receiver<Result<R, E>> where
F: FnOnce(MutexGuard<T>) -> B + 'static,
B: IntoFuture<Item = R, Error = E> + 'static,
R: 'static,
E: 'static,

Acquires a Mutex and performs a computation on its guarded value in a separate task. Returns a Future containing the result of the computation.

This method requires Futures-locks to be build with the "tokio" feature.

When using Tokio, this method will often hold the Mutex for less time than chaining a computation to lock. The reason is that Tokio polls all tasks promptly upon notification. However, Tokio does not guarantee that it will poll all futures promptly when their owning task gets notified. So it's best to hold Mutexes within their own tasks, lest their continuations get blocked by slow stacked combinators.

Examples

let mtx = Mutex::<u32>::new(0);
let r = current_thread::block_on_all(lazy(|| {
    mtx.with(|mut guard| {
        *guard += 5;
        Ok(()) as Result<(), ()>
    })
    .map(|_| assert_eq!(mtx.try_unwrap().unwrap(), 5))
}));
assert!(r.is_ok());