Crate futures_locks

A library of Futures-aware locking primitives. These locks can safely be used in asynchronous environments like Tokio. When they block, they'll only block a single task, not the entire reactor.

Structs

Mutex
A Futures-aware Mutex.

MutexFut
A Future representation a pending Mutex acquisition.

MutexGuard
An RAII mutex guard, much like std::sync::MutexGuard. The wrapped data can be accessed via its Deref and DerefMut implementations.

RwLock
A Futures-aware RwLock.

RwLockReadFut
A Future representation a pending RwLock shared acquisition.

RwLockReadGuard
An RAII guard, much like std::sync::RwLockReadGuard. The wrapped data can be accessed via its Deref implementation.

RwLockWriteFut
A Future representation a pending RwLock exclusive acquisition.

RwLockWriteGuard
An RAII guard, much like std::sync::RwLockWriteGuard. The wrapped data can be accessed via its Deref and DerefMut implementations.